+++ CHANGELOG.md
@@ -4,6 +4,10 @@ PyCheval adheres to [semantic versioning](https://semver.org/).
 
 ## Unreleased
 
+### Fixed
+
+- Don't generate empty trade party fields.
+
 ## 0.3.2 – 2026-02-13
 
 ### Added